# specs/tsp-output/@typespec/openapi3/generated/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: Comprehensive API for managing beauty salon operations including salons, staff, services, customers, reservations, bookings, and reviews.
  version: 0.0.0
...
paths:
...
  /salons:
    get:
      operationId: SalonOperations_list
      summary: List all salons
      parameters:
        - $ref: '#/components/parameters/Models.PaginationParams.limit'
        - $ref: '#/components/parameters/Models.PaginationParams.offset'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - total
                  - limit
                  - offset
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Models.SalonSummary'
                  total:
                    type: integer
                    format: int32
                  limit:
                    type: integer
                    format: int32
                  offset:
                    type: integer
                    format: int32
      tags:
        - Salons
    post:
      operationId: SalonOperations_create
      summary: Create a new salon
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          headers:
            location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Salon'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Salons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.CreateSalonRequest'
...
  /salons/{id}:
    get:
      operationId: SalonOperations_get
      summary: Get salon by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.SalonId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Salon'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Salons
    put:
      operationId: SalonOperations_update
      summary: Update salon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.SalonId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Salon'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Salons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.UpdateSalonRequest'
    delete:
      operationId: SalonOperations_delete
      summary: Delete salon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.SalonId'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.Error'
      tags:
        - Salons
...
components:
  parameters:
    Models.PaginationParams.limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 20
      explode: false
    Models.PaginationParams.offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 0
      explode: false
  schemas:
    Models.Address:
      type: object
      required:
        - street
        - city
        - state
        - postalCode
        - country
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: Japan
...
    Models.ContactInfo:
      type: object
      required:
        - email
        - phoneNumber
      properties:
        email:
          type: string
        phoneNumber:
          type: string
        alternativePhone:
          type: string
...
    Models.CreateSalonRequest:
      type: object
      required:
        - name
        - description
        - address
        - contactInfo
        - openingHours
      properties:
        name:
          type: string
        description:
          type: string
        address:
          $ref: '#/components/schemas/Models.Address'
        contactInfo:
          $ref: '#/components/schemas/Models.ContactInfo'
        openingHours:
          type: array
          items:
            $ref: '#/components/schemas/Models.OpeningHours'
        imageUrls:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
      description: Salon creation request with required and optional fields
...
    Models.DayOfWeek:
      type: string
      enum:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - sunday
...
    Models.Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    Models.OpeningHours:
      type: object
      required:
        - dayOfWeek
        - openTime
        - closeTime
        - isHoliday
      properties:
        dayOfWeek:
          $ref: '#/components/schemas/Models.DayOfWeek'
        openTime:
          type: string
        closeTime:
          type: string
        isHoliday:
          type: boolean
...
    Models.Salon:
      type: object
      required:
        - id
        - name
        - description
        - address
        - contactInfo
        - openingHours
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Models.SalonId'
        name:
          type: string
        description:
          type: string
        address:
          $ref: '#/components/schemas/Models.Address'
        contactInfo:
          $ref: '#/components/schemas/Models.ContactInfo'
        openingHours:
          type: array
          items:
            $ref: '#/components/schemas/Models.OpeningHours'
        imageUrls:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
    Models.SalonId:
      type: string
    Models.SalonSummary:
      type: object
      required:
        - id
        - name
        - address
      properties:
        id:
          $ref: '#/components/schemas/Models.SalonId'
        name:
          type: string
        address:
          $ref: '#/components/schemas/Models.Address'
        rating:
          type: number
          format: float
        reviewCount:
          type: integer
          format: int32
...
    Models.UpdateSalonRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          $ref: '#/components/schemas/Models.Address'
        contactInfo:
          $ref: '#/components/schemas/Models.ContactInfo'
        openingHours:
          type: array
          items:
            $ref: '#/components/schemas/Models.OpeningHours'
        imageUrls:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
      description: Salon update request with optional fields for partial updates
...
servers:
  - url: https://api.beauty-salon.com
    description: Beauty Salon API
    variables: {}